map: Extract `TaskLayer` component
