+++ app/templates/components/map.hbs
@@ -1,5 +1,7 @@
 <div class="map-wrapper">
-  <div id="map" style="width: 100%; height: {{if (and hasDeviceFilter @task) "70%" "100%"}}"></div>
+  <div id="map" style="width: 100%; height: {{if (and hasDeviceFilter @task) "70%" "100%"}}">
+    <TaskLayer @task={{@task}}/>
+  </div>
   {{#if (and hasDeviceFilter @task)}}
     <div class="scores-wrapper">
       <ScoreTable @task={{@task}}/>
